dialects: (csl, csl_wrapper, csl_stencil) move to dialects folder
