--- xdsl/dialects/csl/__init__.py
@@ -1,3 +1,0 @@
-# TID 251 enforces to not import from those
-# We need to skip it here to allow importing from here instead.
-from .csl import *  # noqa: TID251
