+++ xdsl/dialects/csl_stencil.py
