+++ xdsl/transforms/csl_wrapper_hoist_buffers.py
@@ -1,9 +1,8 @@
 from dataclasses import dataclass
 
 from xdsl.context import Context
-from xdsl.dialects import memref
+from xdsl.dialects import csl, csl_wrapper, memref
 from xdsl.dialects.builtin import ModuleOp
-from xdsl.dialects.csl import csl, csl_wrapper
 from xdsl.passes import ModulePass
 from xdsl.pattern_rewriter import (
     PatternRewriter,
