+++ xdsl/transforms/experimental/stencil_tensorize_z_dimension.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass
 from typing import Any, TypeGuard
 
 from xdsl.context import Context
-from xdsl.dialects import builtin, varith
+from xdsl.dialects import builtin, csl_stencil, varith
 from xdsl.dialects.arith import (
     ConstantOp,
     FloatingPointLikeBinaryOperation,
@@ -19,7 +19,6 @@ from xdsl.dialects.builtin import (
     ShapedType,
     TensorType,
 )
-from xdsl.dialects.csl import csl_stencil
 from xdsl.dialects.experimental import dmp
 from xdsl.dialects.func import FuncOp
 from xdsl.dialects.linalg import FillOp
